Add support for new RcheckPrepend entry in .BBSoptions files to allow injection of an arbitrary string in front of the 'R CMD check' command. For example, by adding
    RcheckPrepend: _R_CHECK_FORCE_SUGGESTS_=0

to the .BBSoptions file of package foo, the command used to check foo during
STAGE4 will be '_R_CHECK_FORCE_SUGGESTS_=0 path/to/R CMD check ...'. This can
be used to allow a package to try to pass 'R CMD check' even if some of the
packages it Suggests couldn't be installed (e.g. VariantTools suggests gmapR
but the latter is not supported on Windows). The old way of handling this was
to use Enhances instead of Suggests but this always felt like an ugly hack.