Prerun script now tries to recover when updating a git clone fails
The prerun script now tries to re-create a git clone from scratch on the
central build node if it fails to update it. Updating a git clone can
sometimes fail e.g. if its origin (located on git.bioconductor.org) got
reset. This used to be a fatal error for the builds, unless someone goes
on the build machines to manually delete the "broken" git clones before
the prerun script starts. With this change, we won't need to do this
anymore.