Add packages with invalid 'Package' field in DESCRIPTION to skipped-index.dcf
skipped-index.dcf was implemented to catch packages with a malformed
or missing DESCRIPTION. This situation doesn't fit either case but
but the DESCRIPTION is invalid as per the BBS build standards.
Adding the package to skipped-index.dcf allows the package to appear
on the build report with 'ERROR (Bad DESCRIPTION file)'.