Try to speed up BBS-report.py (and make it more reliable)
BBS-report.py is typically run on the central builder which is where the
build products are. In that case the script will now access these products
via local file access instead of HTTP. This should speed up the reading of
the products and make it more reliable. Using HTTP for this never made much
sense.