Don't display BUILD BIN column on build report when stage is not run
The BUILD BIN column was displayed on the build reports for the
data-annotation and data-experiment subbuilds even though it will
always be empty. Note that even if we decide to run these subbuilds
on Windows or Mac in the future, it's very unlikely that we will run
the BUILD BIN stage.