Small update to Doc/GPU-enabled-builds.md
Make sure to do

    unset CC CXX F77 F90 FC

**after** doing

    module load nvhpc/24.7/nvhpc

This prevents the NVIDIA compiler from becoming the default compiler.
Note that the NVIDIA compiler is not a substitute for gcc in general
e.g. it cannot be used to compile R or Python or any code that relies
on compilation flags supported by the latter but not the former.